[ENG-990] Refactor dispense order view's footer to a component (#16772)
